% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_to_c.r
\name{f_to_c}
\alias{f_to_c}
\title{Fahrenheid converter}
\usage{
f_to_c(temp_F)
}
\arguments{
\item{temp_F}{Numeric value for temperature in Fahrenheid.}
}
\value{
Numeric value representing the temperature in \code{temp_F} converted to Celsius.
}
\description{
This function coverts the input from Fahrenheid to Celsius
}
\examples{
f_to_c(72)
f_to_c(222)
}
\keyword{temperature}
