% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_iris.r
\docType{data}
\name{my_iris}
\alias{my_iris}
\title{My Iris dataset}
\format{
An object of class \code{data.frame} with 150 rows and 5 columns.
}
\source{
Fisher R.A. (1936) he use of multiple measurements in taxonomic problems.
\emph{Annals of Eugenics} 7, Part II, 179–188.
The data were collected by Anderson, Edgar (1935).
The irises of the Gaspe Peninsula, \emph{Bulletin of the American Iris Society}, 59, 2–5.
}
\usage{
my_iris
}
\description{
This data based on iris to be used as example.
}
\keyword{datasets}
