% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pow.r
\name{my_pow}
\alias{my_pow}
\title{Power Function}
\usage{
my_pow(x, power = 2)
}
\arguments{
\item{x}{Numeric value to be raise to the power of \code{power}.}

\item{power}{Numeric value to for the power that \code{x} will raised to
defaults to \code{2}.}
}
\value{
Numeric value representing \code{x} raised to the power \code{power}.
}
\description{
This function raises input to a power.
}
\examples{
my_pow(4)
my_pow(4,3)
}
